演示flink中自定义数据源

+++ src/main/java/com/flink/apps/vo/SenSorReadingV1.java
@@ -28,5 +28,5 @@ public class SenSorReadingV1 implements Serializable {
     /**
      * 温度.
      */
-    private double temperature;
+    private Integer temperature;
 }
